+++ lib/uses.js
@@ -4,6 +4,7 @@ const yaml = require("js-yaml");
 const path = require("path");
 const downloadGitRepo = require("download-git-repo");
 const { spawn } = require("child_process");
+const { handleOutputData } = require("../lib/utils");
 
 /*::
 type ActionYaml = {
@@ -30,19 +31,6 @@ const ensureUses = async (name, target) => {
   // flow-next-uncovered-line
   const res /*:ActionYaml*/ = /*::(*/ mixed /*:any)*/;
   return res;
-};
-
-const countInstances = (rx, text /*:string*/) => {
-  let num = 0;
-  text.replace(rx, () => {
-    num += 1;
-    return "";
-  });
-  return num;
-};
-
-const indent = (text /*: string*/) => {
-  return "|  " + text.replace(/\n(?!$)/g, "\n|  ");
 };
 
 const runUses = async (
@@ -80,31 +68,7 @@ const runUses = async (
       env
     });
 
-    let errors = 0;
-    const errorRx = /^:error:/gm;
-
-    proc.stdout.on("data", (data /*:Buffer*/) => {
-      const text = data.toString("utf8");
-      errors += countInstances(errorRx, text);
-      process.stdout.write(indent(text));
-    });
-
-    proc.stderr.on("data", (data /*:Buffer*/) => {
-      const text = data.toString("utf8");
-      errors += countInstances(errorRx, text);
-      // We use stdout for the stderr as well to avoid interleaving
-      // issues.
-      process.stdout.write(indent(text));
-    });
-
-    proc.on("close", (code /*:number*/) => {
-      if (code === 0) {
-        resolve({ errors, failed: false });
-      } else {
-        console.log(`child process exited with code ${code}`);
-        resolve({ errors, failed: true });
-      }
-    });
+    handleOutputData(proc, resolve);
   });
 };
