+++ lib/utils.js
@@ -1,53 +1,46 @@
 // @flow
 
-// flow-next-uncovered-line
-const prompts = require("prompts");
-const chalk = require("chalk");
-
-const bail = (message /*:string*/ = "Aborting") => {
-  console.error(chalk.red(message));
-  process.exit(1);
-};
-
-const maybeBail = async (message /*:string*/) => {
-  console.error(message);
-  const response = await confirm({ message: "Continue?" });
-  if (!response) {
-    console.error("Aborting");
-    process.exit(1);
-  }
-};
-
-const gotIt = async (message /*:string*/) => {
-  console.log(message);
-  // flow-next-uncovered-line
-  await prompts({ type: "text", name: "ignored", message: "Got it!" });
-};
-
-function prom /*:: <T>*/(
-  fn /*:
-  ((err: ?Error, value: T) => void) => void
-  */
-) /*:Promise<T>*/ {
-  return new Promise((res, rej) =>
-    fn((err, value) => (err ? rej(err) : res(value)))
-  );
-}
-
-const confirm = async (
-  { message, pre } /*: {message: string, pre?: string}*/
-) => {
-  if (pre) {
-    console.error(pre);
-  }
-  /* flow-uncovered-block */
-  const response /*:{continue: boolean}*/ = await prompts({
-    type: "confirm",
-    name: "continue",
-    message
+const countInstances = (rx, text /*:string*/) => {
+  let num = 0;
+  text.replace(rx, () => {
+    num += 1;
+    return "";
   });
-  /* end flow-uncovered-block */
-  return response.continue;
+  return num;
 };
 
-module.exports = { prom, confirm, bail, maybeBail, gotIt };
+const indent = (text /*: string*/) => {
+  return "|  " + text.replace(/\n(?!$)/g, "\n|  ");
+};
+
+const handleOutputData = (proc, resolve) => {
+  let errors = 0;
+  const errorRx = /^:error:/gm;
+
+  proc.stdout.on("data", (data /*:Buffer*/) => {
+    const text = data.toString("utf8");
+    errors += countInstances(errorRx, text);
+    process.stdout.write(indent(text));
+  });
+
+  proc.stderr.on("data", (data /*:Buffer*/) => {
+    const text = data.toString("utf8");
+    errors += countInstances(errorRx, text);
+    // We use stdout for the stderr as well to avoid interleaving
+    // issues.
+    process.stdout.write(indent(text));
+  });
+
+  proc.on("close", (code /*:number*/) => {
+    if (code === 0) {
+      resolve({ errors, failed: false });
+    } else {
+      console.log(`child process exited with code ${code}`);
+      resolve({ errors, failed: true });
+    }
+  });
+};
+
+module.exports = {
+  /*prom, confirm, bail, maybeBail, gotIt,*/ handleOutputData
+};
